Move concurrency helpers
