+++ internal/concurrency/common.go
@@ -1,4 +1,4 @@
-package common
+package concurrency
 
 import (
 	"context"
